demon v 2

+++ New work/demon_v2.py
@@ -4,9 +4,10 @@ import time
 from datetime import datetime
 from random import randint
 import signal
+import psutil
 
 
-PID_FILE = ""
+PID_FILE = "/var/run/step/demon/demon.pid"
 
 def demon():
 
@@ -16,29 +17,64 @@ def demon():
 
     signal.signal(signal.SIGUSR1, handler)
     signal.signal(signal.SIGUSR2, handler)
-    signal.signal(signal.SIGUSR3, handler)
 
     while True:
-        with open("demon.log", "a") as log_file:
-            log_file.write(f"{datetime.now()}\n\n")
-        exept KeyboardInterrupt:
-        with open("signal.log", "a") as log_file:
-            log_file.write("Ctrl + C")
-            exept Exception as e:
+        try:
+            with open("demon.log", "a") as log_file:
+                log_file.write(f"{datetime.now()}\n\n")
+        except KeyboardInterrupt:
+            with open("signal.log", "a") as log_file:
+                log_file.write("Ctrl + C\n")
+        except Exception as e:
             pass
         finally:
             time.sleep(randint(5, 15))
 
 def start_demon():
-    pid = demon()
+    if os.path.isfile(PID_FILE):
+        with open(PID_FILE, "r") as pid_file:
+            pid = int(pid_file.readline())
+            for process in psutil.process_iter():
+                if process.pid == pid:
+                    print("Demon is working.")
+                    return
+    pid = os.fork()
+    if pid:
+        with open(PID_FILE, "w") as pid_file:
+            pid_file.write(f"{pid}")
+        print("Demon was started.")
+        print("Demon has pid: {pid}")
+    else:
+        demon()
 
 def send_signal(args):
     pass
 
+def is_pid_file():
+    return False
+
+# def is_pid_dir():
+#     return False
+
 if __name__ == '__main__':
     try:
-        args = sys.argv[1:]
-    except IndexError:
-        start_demon()
-    else:
+        os.mkdir(os.path.join(*os.path.split(PID_FILE)[:-1]))
+    # except PermissionError:
+    #     Password = input("Введите пароль - ")
+    #     command = f"os.mkdir {os.path.join(*os.path.split(PID_FILE)[:-1])}"
+    #     os.system(f"echo {Password} | sudo -S {command}")
+    #     os.exit(1)
+    except FileNotFoundError:
+        try:
+            os.makedirs(os.path.join(*os.path.split(PID_FILE)[:-1]))
+        except PermissionError:
+            print("WTF !!!!")
+            os.exit(1)
+    except FileExistsError:
+        pass
+    # создавать директорию
+    args = sys.argv[1:]
+    if len(args):
         send_signal(args)
+    else:
+        start_demon()
